refactor: move logic from routes to SellerController
